Created settings file to setup paths and names.
This settings are in the dwl_config.tres file.
Created a function in DWL to pre-load assets from a scene without it having to enter the tree.
DWL registers assets that were already loaded so they are not downloaded again.
Created DWLCustoms.gd for developers to define rules to change the name of the file to load (e.g. if you want to load a different file for mobile devices).