JSON can now register grandchilds.
This will allow the plugin to load not only the assets for a node but also for those grandchilds that are .tscn without needing to go over the node's tree.
Now textures for Control>NinePatchRect and Spatial>MeshInstance can be stored/readed from the JSON.
Nodes that share an asset that is being downloaded get the texture assigned when it is ready.